report

+++ database/panasonic.sql
@@ -1,16 +1,20 @@
 -- phpMyAdmin SQL Dump
--- version 4.7.7
--- https://www.phpmyadmin.net/
+-- version 4.5.1
+-- http://www.phpmyadmin.net
 --
--- Host: localhost:8889
--- Generation Time: Feb 23, 2021 at 12:47 PM
--- Server version: 5.6.38
--- PHP Version: 7.2.1
+-- Host: 127.0.0.1
+-- Generation Time: Feb 24, 2021 at 04:59 PM
+-- Server version: 10.1.10-MariaDB
+-- PHP Version: 5.5.30
 
-SET SQL_MODE
-= "NO_AUTO_VALUE_ON_ZERO";
-SET time_zone
-= "+00:00";
+SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
+SET time_zone = "+00:00";
+
+
+/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
+/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
+/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
+/*!40101 SET NAMES utf8mb4 */;
 
 --
 -- Database: `panasonic`
@@ -22,56 +26,30 @@ SET time_zone
 -- Table structure for table `bill`
 --
 
-CREATE TABLE `bill`
-(
-  `B_id` int
-(11) NOT NULL,
-  `bill_no` varchar
-(20) NOT NULL,
-  `shop` varchar
-(100) NOT NULL,
-  `b_date` varchar
-(100) NOT NULL,
-  `month` int
-(5) NOT NULL,
-  `year` int
-(5) NOT NULL,
-  `bill_amount` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `discount` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `discounted_amt` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `cost` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `cash` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `credit` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `cheque` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `cheque_no` varchar
-(25) NOT NULL,
-  `cheque_date` varchar
-(100) NOT NULL,
-  `cheque_status` smallint
-(6) NOT NULL
+CREATE TABLE `bill` (
+  `B_id` int(11) NOT NULL,
+  `bill_no` varchar(20) NOT NULL,
+  `shop` varchar(100) NOT NULL,
+  `b_date` varchar(100) NOT NULL,
+  `month` int(5) NOT NULL,
+  `year` int(5) NOT NULL,
+  `bill_amount` double(10,2) NOT NULL DEFAULT '0.00',
+  `discount` double(10,2) NOT NULL DEFAULT '0.00',
+  `discounted_amt` double(10,2) NOT NULL DEFAULT '0.00',
+  `cost` double(10,2) NOT NULL DEFAULT '0.00',
+  `cash` double(10,2) NOT NULL DEFAULT '0.00',
+  `credit` double(10,2) NOT NULL DEFAULT '0.00',
+  `cheque` double(10,2) NOT NULL DEFAULT '0.00',
+  `cheque_no` varchar(25) NOT NULL,
+  `cheque_date` varchar(100) NOT NULL,
+  `cheque_status` smallint(6) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 --
 -- Dumping data for table `bill`
 --
 
-INSERT INTO `bill` (`
-B_id`,
-`bill_no`,
-`shop`,
-`b_date`,
-`month`,
-`year`,
-`bill_amount`,
-`discount
-`, `discounted_amt`, `cost`, `cash`, `credit`, `cheque`, `cheque_no`, `cheque_date`, `cheque_status`) VALUES
+INSERT INTO `bill` (`B_id`, `bill_no`, `shop`, `b_date`, `month`, `year`, `bill_amount`, `discount`, `discounted_amt`, `cost`, `cash`, `credit`, `cheque`, `cheque_no`, `cheque_date`, `cheque_status`) VALUES
 (4, '00000004', 'shop 3', '2021-01-12', 1, 2021, 350.00, 0.00, 350.00, 250.00, 150.00, 250.00, 0.00, '', '', 0),
 (5, '00000003', 'shop 4', '2021-01-12', 1, 2021, 1500.00, 150.00, 1350.00, 500.00, 1350.00, 0.00, 0.00, '', '', 0),
 (6, '00000005', 'Shop 3', '2021-01-13', 1, 2021, 6480.00, 0.00, 6480.00, 5520.00, 0.00, 1480.00, 5000.00, '125478459', '2021-02-06', 1),
@@ -91,10 +69,8 @@ B_id`,
 -- Table structure for table `category`
 --
 
-CREATE TABLE `category`
-(
-  `category_id` int
-(11) NOT NULL,
+CREATE TABLE `category` (
+  `category_id` int(11) NOT NULL,
   `category_name` text NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
@@ -102,10 +78,7 @@ CREATE TABLE `category`
 -- Dumping data for table `category`
 --
 
-INSERT INTO `category` (`
-category_id`,
-`category_name
-`) VALUES
+INSERT INTO `category` (`category_id`, `category_name`) VALUES
 (2, 'MANGANESE'),
 (3, 'ALKALINE'),
 (4, 'EVOLTA'),
@@ -119,36 +92,23 @@ category_id`,
 -- Table structure for table `credit`
 --
 
-CREATE TABLE `credit`
-(
-  `c_id` int
-(11) NOT NULL,
-  `invoice_no` varchar
-(25) NOT NULL,
-  `cdate` varchar
-(100) NOT NULL,
-  `month` int
-(5) NOT NULL,
-  `year` int
-(5) NOT NULL,
-  `amount` double
-(10,2) NOT NULL,
-  `type` varchar
-(10) NOT NULL,
-  `remain` double
-(10,2) NOT NULL,
-  `credit_status` smallint
-(1) NOT NULL
+CREATE TABLE `credit` (
+  `c_id` int(11) NOT NULL,
+  `invoice_no` varchar(25) NOT NULL,
+  `cdate` varchar(100) NOT NULL,
+  `month` int(5) NOT NULL,
+  `year` int(5) NOT NULL,
+  `amount` double(10,2) NOT NULL,
+  `type` varchar(10) NOT NULL,
+  `remain` double(10,2) NOT NULL,
+  `credit_status` smallint(1) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 --
 -- Dumping data for table `credit`
 --
 
-INSERT INTO `credit` (`
-c_id`,
-`invoice_no
-`, `cdate`, `month`, `year`, `amount`, `type`, `remain`, `credit_status`) VALUES
+INSERT INTO `credit` (`c_id`, `invoice_no`, `cdate`, `month`, `year`, `amount`, `type`, `remain`, `credit_status`) VALUES
 (1, 'PS1001', '2021-01-23', 1, 2021, 12500.00, 'invoice', 12500.00, 0),
 (2, 'PS1001', '2021-01-25', 1, 2021, 2000.00, 'payment', 10500.00, 0),
 (3, 'PS1002', '2021-01-29', 1, 2021, 24000.00, 'invoice', 24000.00, 1),
@@ -164,30 +124,20 @@ c_id`,
 -- Table structure for table `debt`
 --
 
-CREATE TABLE `debt`
-(
-  `d_id` int
-(11) NOT NULL,
-  `shop` varchar
-(100) NOT NULL,
-  `ddate` varchar
-(100) NOT NULL,
-  `month` int
-(5) NOT NULL,
-  `year` int
-(5) NOT NULL,
-  `amt` double
-(10,2) NOT NULL DEFAULT '0.00'
+CREATE TABLE `debt` (
+  `d_id` int(11) NOT NULL,
+  `shop` varchar(100) NOT NULL,
+  `ddate` varchar(100) NOT NULL,
+  `month` int(5) NOT NULL,
+  `year` int(5) NOT NULL,
+  `amt` double(10,2) NOT NULL DEFAULT '0.00'
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 --
 -- Dumping data for table `debt`
 --
 
-INSERT INTO `debt` (`
-d_id`,
-`shop
-`, `ddate`, `month`, `year`, `amt`) VALUES
+INSERT INTO `debt` (`d_id`, `shop`, `ddate`, `month`, `year`, `amt`) VALUES
 (1, 'shop 3', '2021-01-16', 1, 2021, 230.00),
 (3, 'shop 3', '2021-01-23', 1, 2021, 250.00),
 (5, 'shop 5', '2021-01-29', 1, 2021, 3000.00),
@@ -201,37 +151,25 @@ d_id`,
 -- Table structure for table `item`
 --
 
-CREATE TABLE `item`
-(
-  `item_id` int
-(11) NOT NULL,
+CREATE TABLE `item` (
+  `item_id` int(11) NOT NULL,
   `category` text NOT NULL,
   `item_name` text NOT NULL,
   `batch_no` text NOT NULL,
   `size` text NOT NULL,
-  `purchase_cost` double
-(10,2) NOT NULL,
-  `sales_cost` double
-(10,2) NOT NULL,
-  `warehouse_stock` int
-(11) NOT NULL,
-  `free_stock` int
-(11) NOT NULL,
-  `lorry_stock` int
-(11) NOT NULL,
-  `lorry_free_stock` int
-(11) NOT NULL
+  `purchase_cost` double(10,2) NOT NULL,
+  `sales_cost` double(10,2) NOT NULL,
+  `warehouse_stock` int(11) NOT NULL,
+  `free_stock` int(11) NOT NULL,
+  `lorry_stock` int(11) NOT NULL,
+  `lorry_free_stock` int(11) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 --
 -- Dumping data for table `item`
 --
 
-INSERT INTO `item` (`
-item_id`,
-`category`,
-`item_name
-`, `batch_no`, `size`, `purchase_cost`, `sales_cost`, `warehouse_stock`, `free_stock`, `lorry_stock`, `lorry_free_stock`) VALUES
+INSERT INTO `item` (`item_id`, `category`, `item_name`, `batch_no`, `size`, `purchase_cost`, `sales_cost`, `warehouse_stock`, `free_stock`, `lorry_stock`, `lorry_free_stock`) VALUES
 (1, 'MANGANESE', 'R6NT/1B12', '1', 'AA', 100.00, 110.00, 745, 60, 282, 6),
 (2, 'CHARGERS', 'R6NT/1B123S', '3', 'AAAA', 160.00, 180.00, 670, 25, 64, 7),
 (3, 'ALKALINE', 'LR6T/4B', '2', 'AAA', 100.00, 125.00, 520, 40, 199, 1),
@@ -244,30 +182,20 @@ item_id`,
 -- Table structure for table `profit`
 --
 
-CREATE TABLE `profit`
-(
-  `P_id` int
-(11) NOT NULL,
-  `cdate` varchar
-(100) NOT NULL,
-  `sales` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `purchase_cost` double
-(10,2) NOT NULL,
-  `expenses` double
-(10,2) NOT NULL DEFAULT '0.00',
-  `daily_profit` double
-(10,2) NOT NULL DEFAULT '0.00'
+CREATE TABLE `profit` (
+  `P_id` int(11) NOT NULL,
+  `cdate` varchar(100) NOT NULL,
+  `sales` double(10,2) NOT NULL DEFAULT '0.00',
+  `purchase_cost` double(10,2) NOT NULL,
+  `expenses` double(10,2) NOT NULL DEFAULT '0.00',
+  `daily_profit` double(10,2) NOT NULL DEFAULT '0.00'
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 --
 -- Dumping data for table `profit`
 --
 
-INSERT INTO `profit` (`
-P_id`,
-`cdate
-`, `sales`, `purchase_cost`, `expenses`, `daily_profit`) VALUES
+INSERT INTO `profit` (`P_id`, `cdate`, `sales`, `purchase_cost`, `expenses`, `daily_profit`) VALUES
 (1, '2021-01-12', 1700.00, 750.00, 250.00, 700.00),
 (3, '2021-01-13', 7800.00, 6720.00, 220.00, 860.00),
 (4, '2021-01-24', 1320.00, 1200.00, 0.00, 120.00);
@@ -278,36 +206,23 @@ P_id`,
 -- Table structure for table `sale_items`
 --
 
-CREATE TABLE `sale_items`
-(
-  `sale_id` int
-(11) NOT NULL,
-  `bill_no` varchar
-(20) NOT NULL,
-  `item` varchar
-(150) NOT NULL,
-  `total` int
-(11) NOT NULL DEFAULT '0',
-  `sale` int
-(11) NOT NULL DEFAULT '0',
-  `free` int
-(11) NOT NULL DEFAULT '0',
-  `af_bal` int
-(11) NOT NULL DEFAULT '0',
-  `total_free` int
-(11) NOT NULL,
-  `af_free` int
-(11) NOT NULL
+CREATE TABLE `sale_items` (
+  `sale_id` int(11) NOT NULL,
+  `bill_no` varchar(20) NOT NULL,
+  `item` varchar(150) NOT NULL,
+  `total` int(11) NOT NULL DEFAULT '0',
+  `sale` int(11) NOT NULL DEFAULT '0',
+  `free` int(11) NOT NULL DEFAULT '0',
+  `af_bal` int(11) NOT NULL DEFAULT '0',
+  `total_free` int(11) NOT NULL,
+  `af_free` int(11) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 --
 -- Dumping data for table `sale_items`
 --
 
-INSERT INTO `sale_items` (`
-sale_id`,
-`bill_no
-`, `item`, `total`, `sale`, `free`, `af_bal`, `total_free`, `af_free`) VALUES
+INSERT INTO `sale_items` (`sale_id`, `bill_no`, `item`, `total`, `sale`, `free`, `af_bal`, `total_free`, `af_free`) VALUES
 (6, '00000004', 'R6NT/1B12', 150, 12, 3, 135, 0, 0),
 (7, '00000003', 'R6NT/1B123S', 83, 12, 3, 68, 0, 0),
 (8, '00000005', 'LR6T/4B', 50, 24, 0, 26, 0, 0),
@@ -336,38 +251,25 @@ sale_id`,
 -- Table structure for table `trxn`
 --
 
-CREATE TABLE `trxn`
-(
-  `trxn_id` int
-(11) NOT NULL,
+CREATE TABLE `trxn` (
+  `trxn_id` int(11) NOT NULL,
   `category` text NOT NULL,
   `item` text NOT NULL,
   `size` text NOT NULL,
-  `load_bal` int
-(11) NOT NULL,
-  `free_load` int
-(11) NOT NULL,
-  `bf_bal` int
-(11) NOT NULL,
-  `bf_free` int
-(11) NOT NULL,
-  `total` int
-(11) NOT NULL,
-  `tot_free` int
-(11) NOT NULL,
-  `create_date` varchar
-(100) NOT NULL
+  `load_bal` int(11) NOT NULL,
+  `free_load` int(11) NOT NULL,
+  `bf_bal` int(11) NOT NULL,
+  `bf_free` int(11) NOT NULL,
+  `total` int(11) NOT NULL,
+  `tot_free` int(11) NOT NULL,
+  `create_date` varchar(100) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 --
 -- Dumping data for table `trxn`
 --
 
-INSERT INTO `trxn` (`
-trxn_id`,
-`category`,
-`item
-`, `size`, `load_bal`, `free_load`, `bf_bal`, `bf_free`, `total`, `tot_free`, `create_date`) VALUES
+INSERT INTO `trxn` (`trxn_id`, `category`, `item`, `size`, `load_bal`, `free_load`, `bf_bal`, `bf_free`, `total`, `tot_free`, `create_date`) VALUES
 (1, 'MANGANESE', 'R6NT/1B12', 'AA', 25, 0, 100, 0, 125, 0, '2021-01-02'),
 (2, 'CHARGERS', 'R6NT/1B123S', 'AAAA', 70, 0, 0, 0, 70, 0, '2021-01-02'),
 (3, 'ALKALINE', 'LR6T/4B', 'AAA', 50, 0, 0, 0, 50, 0, '2021-01-02'),
@@ -394,25 +296,18 @@ trxn_id`,
 -- Table structure for table `user`
 --
 
-CREATE TABLE `user`
-(
-  `user_id` int
-(11) NOT NULL,
-  `username` varchar
-(50) NOT NULL,
-  `password` varchar
-(50) NOT NULL,
-  `role` int
-(11) NOT NULL
+CREATE TABLE `user` (
+  `user_id` int(11) NOT NULL,
+  `username` varchar(50) NOT NULL,
+  `password` varchar(50) NOT NULL,
+  `role` int(11) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
 --
 -- Dumping data for table `user`
 --
 
-INSERT INTO `user`
-  (`user_id`, `username
-`, `password`, `role`) VALUES
+INSERT INTO `user` (`user_id`, `username`, `password`, `role`) VALUES
 (1, 'admin', '21232f297a57a5a743894a0e4a801fc3', 1),
 (2, 'user', 'ee11cbb19052e40b07aac0ca060c23ee', 2);
 
@@ -424,64 +319,55 @@ INSERT INTO `user`
 -- Indexes for table `bill`
 --
 ALTER TABLE `bill`
-ADD PRIMARY KEY
-(`B_id`);
+  ADD PRIMARY KEY (`B_id`);
 
 --
 -- Indexes for table `category`
 --
 ALTER TABLE `category`
-ADD PRIMARY KEY
-(`category_id`);
+  ADD PRIMARY KEY (`category_id`);
 
 --
 -- Indexes for table `credit`
 --
 ALTER TABLE `credit`
-ADD PRIMARY KEY
-(`c_id`);
+  ADD PRIMARY KEY (`c_id`);
 
 --
 -- Indexes for table `debt`
 --
 ALTER TABLE `debt`
-ADD PRIMARY KEY
-(`d_id`);
+  ADD PRIMARY KEY (`d_id`);
 
 --
 -- Indexes for table `item`
 --
 ALTER TABLE `item`
-ADD PRIMARY KEY
-(`item_id`);
+  ADD PRIMARY KEY (`item_id`);
 
 --
 -- Indexes for table `profit`
 --
 ALTER TABLE `profit`
-ADD PRIMARY KEY
-(`P_id`);
+  ADD PRIMARY KEY (`P_id`);
 
 --
 -- Indexes for table `sale_items`
 --
 ALTER TABLE `sale_items`
-ADD PRIMARY KEY
-(`sale_id`);
+  ADD PRIMARY KEY (`sale_id`);
 
 --
 -- Indexes for table `trxn`
 --
 ALTER TABLE `trxn`
-ADD PRIMARY KEY
-(`trxn_id`);
+  ADD PRIMARY KEY (`trxn_id`);
 
 --
 -- Indexes for table `user`
 --
 ALTER TABLE `user`
-ADD PRIMARY KEY
-(`user_id`);
+  ADD PRIMARY KEY (`user_id`);
 
 --
 -- AUTO_INCREMENT for dumped tables
@@ -491,61 +377,47 @@ ADD PRIMARY KEY
 -- AUTO_INCREMENT for table `bill`
 --
 ALTER TABLE `bill`
-  MODIFY `B_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=16;
-
+  MODIFY `B_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=16;
 --
 -- AUTO_INCREMENT for table `category`
 --
 ALTER TABLE `category`
-  MODIFY `category_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=8;
-
+  MODIFY `category_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=8;
 --
 -- AUTO_INCREMENT for table `credit`
 --
 ALTER TABLE `credit`
-  MODIFY `c_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=9;
-
+  MODIFY `c_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=9;
 --
 -- AUTO_INCREMENT for table `debt`
 --
 ALTER TABLE `debt`
-  MODIFY `d_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=9;
-
+  MODIFY `d_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=9;
 --
 -- AUTO_INCREMENT for table `item`
 --
 ALTER TABLE `item`
-  MODIFY `item_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=6;
-
+  MODIFY `item_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=6;
 --
 -- AUTO_INCREMENT for table `profit`
 --
 ALTER TABLE `profit`
-  MODIFY `P_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=5;
-
+  MODIFY `P_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=5;
 --
 -- AUTO_INCREMENT for table `sale_items`
 --
 ALTER TABLE `sale_items`
-  MODIFY `sale_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=27;
-
+  MODIFY `sale_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=27;
 --
 -- AUTO_INCREMENT for table `trxn`
 --
 ALTER TABLE `trxn`
-  MODIFY `trxn_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=25;
-
+  MODIFY `trxn_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=25;
 --
 -- AUTO_INCREMENT for table `user`
 --
 ALTER TABLE `user`
-  MODIFY `user_id` int
-(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=3;
+  MODIFY `user_id` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=3;
+/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
+/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
+/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
